+++ interface/build.gradle
@@ -7,8 +7,6 @@ apply from: 'publish.gradle'
 dependencies {
   implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
-  api "io.reactivex.rxjava2:rxjava:$rxJavaVersion"
-
   testImplementation 'junit:junit:4.12'
   testImplementation 'org.mockito:mockito-inline:2.23.0'
   testImplementation 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.1.0'
